Also delete topics which 404 as well as ones with 0 subscribers

+++ lib/data_hygiene/delete_unneeded_topics.rb
@@ -1,7 +1,7 @@
 require 'thread'
 
 module DataHygiene
-  class DeleteTopicsWithoutSubscribers
+  class DeleteUnneededTopics
     attr_reader :client
     delegate :puts, to: :@output
     delegate :gets, to: :@input
@@ -15,20 +15,30 @@ module DataHygiene
     def call
       puts "#{with_zero_subscribers.count} subscriber lists without subscribers"
       with_zero_subscribers.each { |sl| puts "#{sl.gov_delivery_id} - #{sl.title}"}
+      if with_missing_topic.count > 0
+        puts ''
+        puts "#{with_missing_topic.count} subscriber lists don't exist in GovDelivery"
+        with_missing_topic.each { |sl| puts "#{sl.gov_delivery_id} - #{sl.title}"}
+      end
       if with_gov_delivery_error.count > 0
         puts ''
         puts "#{with_gov_delivery_error.count} subscriber lists with errors - THESE ARE NOT BEING DELETED"
         with_gov_delivery_error.each { |sl| puts "#{sl.gov_delivery_id} - #{sl.title}"}
       end
 
-      puts ''
-      puts "#{with_zero_subscribers.count} subscriber lists without subscribers, enter `delete` to delete from database and GovDelivery: "
-      command = gets
+      to_delete = with_zero_subscribers + with_missing_topic
+      if to_delete.count > 0
+        puts ''
+        puts "#{with_zero_subscribers.count} subscriber lists without subscribers and #{with_missing_topic.count} subscriber lists don't exist in GovDelivery, enter `delete` to delete from database and GovDelivery: "
+        command = gets
 
-      if command.chomp == 'delete'
-        with_zero_subscribers.each do |subscriber_list|
-          delete_topic(subscriber_list)
+        if command.chomp == 'delete'
+          to_delete.each do |subscriber_list|
+            delete_topic(subscriber_list)
+          end
         end
+      else
+        puts 'Nothing to delete'
       end
     end
 
@@ -38,13 +48,19 @@ module DataHygiene
 
     def with_zero_subscribers
       @with_zero_subscribers ||= subscriber_lists
-        .select { |c, s| c&.zero? }
+        .select { |c, s| c == '0' }
+        .map(&:last)
+    end
+
+    def with_missing_topic
+      @with_missing_topic ||= subscriber_lists
+        .select { |c, s| c == :topic_not_found }
         .map(&:last)
     end
 
     def with_gov_delivery_error
       @with_gov_delivery_error ||= subscriber_lists
-        .select { |c, s| c.nil? }
+        .select { |c, s| c == :unknown_error }
         .map(&:last)
     end
 
@@ -52,20 +68,27 @@ module DataHygiene
 
     def subscriber_count(gov_delivery_id)
       topic = client.fetch_topic(gov_delivery_id)
-      topic && topic['subscribers_count']
+      topic.subscribers_count
+    rescue GovDelivery::Client::TopicNotFound
+      :topic_not_found
     rescue GovDelivery::Client::UnknownError
-      nil
+      :unknown_error
     end
 
     def delete_topic(subscriber_list)
       subscribers = subscriber_count(subscriber_list.gov_delivery_id)
 
-      if subscribers.nil?
-        puts "Skipping #{subscriber_list.gov_delivery_id} as it does not appear on GovDelivery"
-      elsif subscribers.zero?
+      case subscribers
+      when '0'
         puts "Deleting: #{subscriber_list.gov_delivery_id}"
-        delete_topic_with_error_catching(subscriber_list.gov_delivery_id)
+        if delete_topic_with_error_catching(subscriber_list.gov_delivery_id)
+          SubscriberList.where(gov_delivery_id: subscriber_list.gov_delivery_id).delete_all
+        end
+      when :topic_not_found
+        puts "Deleting: #{subscriber_list.gov_delivery_id}"
         SubscriberList.where(gov_delivery_id: subscriber_list.gov_delivery_id).delete_all
+      when :unknown_error
+        puts "Skipping #{subscriber_list.gov_delivery_id} as it does not appear on GovDelivery"
       else
         puts "Skipping #{subscriber_list.gov_delivery_id} as it now has #{subscribers} subscribers"
       end
@@ -73,8 +96,10 @@ module DataHygiene
 
     def delete_topic_with_error_catching(gov_delivery_id)
       client.delete_topic(gov_delivery_id)
+      true
     rescue GovDelivery::Client::UnknownError
       puts "---- Error deleting: #{gov_delivery_id}"
+      false
     end
 
     class ThreadedSubscriberCountRetriever
@@ -115,9 +140,11 @@ module DataHygiene
 
       def subscriber_count(gov_delivery_id)
         topic = client.fetch_topic(gov_delivery_id)
-        topic && topic['subscribers_count']
+        topic.subscribers_count
+      rescue GovDelivery::Client::TopicNotFound
+        :topic_not_found
       rescue GovDelivery::Client::UnknownError
-        nil
+        :unknown_error
       end
     end
   end
